requirements.txt
click==6.7
requests==2.12.4
Fabric3==1.13.1.post1
python_dateutil==2.6.1
Fabric==1.13.2
